update server for graph
